Разделила код на модули, переделала функционал генерации id для фото и коментов.
